requirements.txt
Django>=4.2.7,<6.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-filter>=23.3
gunicorn>=21.2.0
whitenoise>=6.5.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
requests>=2.31.0
Pillow>=11.0.0
asgiref>=3.8.1
certifi>=2025.4.26
charset-normalizer>=3.4.2
idna>=3.10
packaging>=25.0
soupsieve>=2.7
sqlparse>=0.5.3
typing_extensions>=4.13.2
tzdata>=2025.2
urllib3>=2.4.0
